+++ src/main/java/org/lyg/vpc/view/UsrToken.java
@@ -8,5 +8,5 @@ public class UsrToken {
     private String uKey;
     private String uPassword;
     private String uToken;
-    private Integer uTime;
+    private long uTime;
 }
